feat: Extract AssociatedOutput unit to own component and DRY code + Support multiple concurrent expanded outputs
